+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Locale data support for asset information
 - Update GraphQL playground
 - Add app:request:after event triggered after fastcgi_finish_request (if available)
+- Improve tower terminal ux (admin ui)
 
 ## 2.10.0 (2024-11-05)